Extract volume to own object and change null checks in thymeleaf
The volume information is used other places, so it's good to pull it
out into its own object. Changed how the null checks are done in the
thymeleaf templates to let thymeleaf handle it, instead of spring. Also
made change to fix a problem where the th:replace has a higher
precedence than th:if, so the th:if was being ignored.